refactor: remove unused options from github-ci-pack-verify.sh and add new compatibility tests for sample extensions
